Add SelectionChanged and TextChanged events that are called from the TextFile after changes are made, these events lead to invalidating of the textfield rather than explicitly invalidating.

+++ PantherGUI/control.cpp
@@ -60,7 +60,7 @@ bool Control::KeyboardUnicode(PGUTF8Character character, PGModifier modifier) {
 }
 
 void Control::Invalidate() {
-	RefreshWindow(this->window, PGIRect(X(), Y(), this->width, this->height));
+	RefreshWindow(this->window, PGIRect((int)X(), (int)Y(), (int)this->width, (int)this->height));
 }
 
 void Control::Invalidate(PGIRect rectangle) {
